A lot of changes, now seems to work as before
