+++ Readme.md
@@ -1,13 +1,9 @@
 # Gaiku (WIP very early development stage)
 
-[2d](gaiku-2d) and [3d](gaiku-3d) agnostic framework (game engine) terrain engine. The purpose of the Gaiku is to provide a 
+3d agnostic framework (game engine) terrain engine. The purpose of the Gaiku is to provide a 
 foundation to build your games easily without worrying about implementing your own terrain engine.
 
 The main crate is developed game engine agnostic, so we can help to reach more developers.
-
-## Supported game engines
-
-- [Amethyst](gaiku-amethyst)
 
 ## Features
 
@@ -23,15 +19,11 @@ General features:
 - Chunk based for infinite generated terrains
 - Mesh optimization
 - LOD support
-- For procedural terrain generation check the 2d and 3d `Bakers` documentations.
-
-For specific features check [2d](gaiku-2d/Readme.md) and [3d](gaiku-3d/Readme.md)
+- For procedural terrain generation check `Bakers` documentations.
 
 ## Examples
 
-### 3d
-
-To run the examples go to the folder `gaiku-3d` and run with:
+To run the examples go to the folder `gaiku` and run with:
 
 ```bash
 cargo run --example heightmap
@@ -61,12 +53,3 @@ Tree
 
 
 * *There are some kown and unresolved issues with the marching cubes baker.*
-
-## License
-
-Licensed under either of
-
-- Apache License, Version 2.0, ([license/APACHE](license/APACHE) or http://www.apache.org/licenses/LICENSE-2.0)
-- MIT license ([license/MIT](license/MIT) or http://opensource.org/licenses/MIT)
-
-at your option.
