+++ crates/gaiku-bevy/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "gaiku-bevy"
+version = "0.1.0"
+authors = ["Norman Paniagua <normanpaniagua@gmail.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+bevy = "0.1.3"
+gaiku = { path = "../../" }
+
+[dev-dependencies]
+bevy_fly_camera = "0.2.0"
+
+[[example]]
+name = "main"
+path = "examples/main.rs"
