+++ crates/gaiku-common/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "gaiku-common"
+version = "0.1.0"
+authors = ["Norman Paniagua <normanpaniagua@gmail.com>"]
+edition = "2018"
+description = "Common library used by gaiku"
+license = "MIT"
+
+documentation = "https://docs.rs/gaiku-common"
+homepage = "https://github.com/norman784/gaiku"
+repository = "https://github.com/norman784/gaiku"
+
+[dependencies]
+decorum = "0.1.3"
+glam = "0.9.4"
+num-traits = "0.2.6"
