+++ crates/gaiku-common/src/data.rs
@@ -1,0 +1,7 @@
+mod chunk;
+mod mesh;
+
+pub use self::{
+  chunk::Chunk,
+  mesh::{Mesh, MeshBuilder},
+};
