+++ examples/assets/heightmap.png
