+++ examples/assets/planet.gox
