+++ examples/assets/small_tree.gox
