+++ examples/assets/terrain.gox
