+++ examples/common/mod.rs
