--- gaiku-2d/example/src/main.rs
@@ -1,3 +1,0 @@
-fn main() {
-  println!("Gaiku 2d");
-}
