--- gaiku-3d/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-name = "gaiku-3d"
-version = "0.1.0"
-authors = ["Norman Paniagua <normanpaniagua@gmail.com>"]
-edition = "2018"
-description = "3d agnostic framework (game engine) terrain engine."
-exclude = ["gaiku-3d/examples/*"]
-keywords = ["3d", "terrain", "engine", "gaiku"]
-categories = ["game-engines"]
-
-documentation = "https://docs.rs/gaiku-3d"
-homepage = "https://github.com/norman784/gaiku"
-repository = "https://github.com/norman784/gaiku"
-
-readme = "Readme.md"
-license = "MIT/Apache-2.0"
-
-[dependencies]
-gaiku-common = { path = "../gaiku-common", version = "0.1.0" }
-gox = "0.2.0"
-#gox = { path = "../../gox" }
-png = "0.14.0"
-
-[dev-dependencies]
-obj-exporter = "0.2.0"
