--- gaiku-amethyst/Cargo.toml
@@ -1,9 +1,0 @@
-[package]
-name = "gaiku-amethyst"
-version = "0.1.0"
-authors = ["Norman Paniagua <normanpaniagua@gmail.com>"]
-edition = "2018"
-
-[dependencies]
-amethyst = "0.10.0"
-gaiku = { path = "gaiku", version = "0.1.0" }
