--- gaiku-amethyst/Readme.md
@@ -1,5 +1,0 @@
-# Gaiku Amethyst
-
-Gaiku implementation for [Amethyst game engine](https://amethyst.rs).
-
-## Example
