--- gaiku-common/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "gaiku-common"
-version = "0.1.0"
-authors = ["Norman Paniagua <normanpaniagua@gmail.com>"]
-edition = "2018"
-description = "Common library used by gaiku"
-license = "MIT/Apache-2.0"
-
-documentation = "https://docs.rs/gaiku-common"
-homepage = "https://github.com/norman784/gaiku"
-repository = "https://github.com/norman784/gaiku"
-
-[dependencies]
-decorum = "0.1.3"
-derive-new = "0.5.6"
-getset = "0.0.7"
-glam = { version = "0.8.5", features = ["mint"] }
-mint = "0.5.4"
-num-traits = "0.2.6"
-typed-builder = "0.4.1"
