--- gaiku-common/src/data/mod.rs
@@ -1,4 +1,0 @@
-mod chunk;
-mod mesh;
-
-pub use self::{chunk::Chunk, mesh::Mesh};
