+++ rustfmt.toml
@@ -1,0 +1,5 @@
+merge_imports = true
+tab_spaces = 2
+use_field_init_shorthand = true
+reoder_impl_items = true
+newline_style = "Unix"
